+++ src/app/login/login.module.css
@@ -1,0 +1,16 @@
+.container {
+  padding-top: 90px;
+  /* justify-content: space-between;
+  flex-direction: column;
+  align-items: center; */
+  height: 100%;
+}
+.wrapper {
+  padding: 0 20px;
+}
+.nav {
+  height: 60px;
+  width: 100%;
+  display: flex;
+  align-items: center;
+}
