--- src/app/signup/signup.module.css
@@ -1,16 +1,0 @@
-.container {
-  padding-top: 146px;
-  display: flex;
-  justify-content: space-between;
-  flex-direction: column;
-  align-items: center;
-  height: 100%;
-}
-.h1 {
-  font-family: Pretendard;
-  font-size: 16px;
-  font-style: normal;
-  font-weight: 500;
-  line-height: 150%; /* 24px */
-  margin-bottom: 8px;
-}
